+++ BetterChests/Services/ManagedChests.cs
@@ -4,13 +4,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Common.Extensions;
-using Common.Helpers;
-using Common.Records;
 using Microsoft.Xna.Framework;
 using StardewModdingAPI;
-using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
-using StardewMods.BetterChests.Extensions;
 using StardewMods.BetterChests.Interfaces;
 using StardewMods.BetterChests.Models;
 using StardewMods.FuryCore.Interfaces;
@@ -23,8 +19,7 @@ using SObject = StardewValley.Object;
 internal class ManagedChests : IModService
 {
     private readonly Lazy<AssetHandler> _assetHandler;
-    private readonly PerScreen<Dictionary<PlacedChest, Lazy<IManagedChest>>> _placedChests = new(() => new());
-    private readonly PerScreen<Dictionary<PlayerItem, IManagedChest>> _playerChests = new(() => new());
+    private readonly PerScreen<IDictionary<Item, IManagedChest>> _cachedObjects = new(() => new Dictionary<Item, IManagedChest>());
 
     /// <summary>
     ///     Initializes a new instance of the <see cref="ManagedChests" /> class.
@@ -37,32 +32,112 @@ internal class ManagedChests : IModService
         this.Config = config;
         this.Helper = helper;
         this._assetHandler = services.Lazy<AssetHandler>();
-        PlacedChestExtensions.GetAccessibleLocations = () => this.AccessibleLocations;
-        this.Helper.Events.GameLoop.SaveLoaded += this.OnSaveLoaded;
-        this.Helper.Events.Player.InventoryChanged += this.OnInventoryChanged;
-        this.Helper.Events.World.ObjectListChanged += this.OnObjectListChanged;
-        this.Helper.Events.Multiplayer.ModMessageReceived += this.OnModMessageReceived;
-
-        if (Context.IsMainPlayer)
-        {
-            this.Helper.Events.Multiplayer.PeerConnected += this.OnPeerConnected;
-        }
     }
 
     /// <summary>
     ///     Gets all placed chests in the world.
     /// </summary>
-    public IReadOnlyDictionary<PlacedChest, Lazy<IManagedChest>> PlacedChests
+    public IEnumerable<KeyValuePair<KeyValuePair<GameLocation, Vector2>, IManagedChest>> PlacedChests
     {
-        get => this._placedChests.Value;
+        get
+        {
+            foreach (var location in this.AccessibleLocations)
+            {
+                // Return fridge if location is FarmHouse
+                if (location is FarmHouse { fridge.Value: { } fridge, fridgePosition: var point } && !point.ToVector2().Equals(Vector2.Zero))
+                {
+                    if (!this.CachedObjects.TryGetValue(fridge, out var managedChest))
+                    {
+                        managedChest = new ManagedChest(fridge, this.GetChestData("Fridge"), "Fridge");
+                        this.CachedObjects.Add(fridge, managedChest);
+                    }
+
+                    if (managedChest is not null)
+                    {
+                        yield return new(new(location, point.ToVector2()), managedChest);
+                    }
+                }
+
+                foreach (var (position, obj) in location.Objects.Pairs)
+                {
+                    // Add untracked objects to cache
+                    if (!this.CachedObjects.TryGetValue(obj, out var managedChest))
+                    {
+                        var chest = obj switch
+                        {
+                            Chest { Stack: 1 } playerChest when playerChest.IsPlayerChest() => playerChest,
+                            { Stack: 1, heldObject.Value: Chest heldChest } when heldChest.IsPlayerChest() => heldChest,
+                            _ => null,
+                        };
+
+                        if (chest is null)
+                        {
+                            this.CachedObjects.Add(obj, null);
+                            continue;
+                        }
+
+                        var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == obj.ParentSheetIndex).Value?[0];
+                        if (string.IsNullOrWhiteSpace(name))
+                        {
+                            this.CachedObjects.Add(obj, null);
+                            continue;
+                        }
+
+                        managedChest = new ManagedChest(chest, this.GetChestData(name), name);
+                        this.CachedObjects.Add(chest, managedChest);
+                    }
+
+                    if (managedChest is not null)
+                    {
+                        yield return new(new(location, position), managedChest);
+                    }
+                }
+            }
+        }
     }
 
     /// <summary>
     ///     Gets all chests in player inventory.
     /// </summary>
-    public IReadOnlyDictionary<PlayerItem, IManagedChest> PlayerChests
+    public IEnumerable<IManagedChest> PlayerChests
     {
-        get => this._playerChests.Value;
+        get
+        {
+            foreach (var item in Game1.player.Items.Where(item => item is not null))
+            {
+                // Add untracked objects to cache
+                if (!this.CachedObjects.TryGetValue(item, out var managedChest))
+                {
+                    var chest = item switch
+                    {
+                        Chest { Stack: 1 } playerChest when playerChest.IsPlayerChest() => playerChest,
+                        SObject { Stack: 1, heldObject.Value: Chest heldChest } when heldChest.IsPlayerChest() => heldChest,
+                        _ => null,
+                    };
+
+                    if (chest is null)
+                    {
+                        this.CachedObjects.Add(item, null);
+                        continue;
+                    }
+
+                    var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == item.ParentSheetIndex).Value?[0];
+                    if (string.IsNullOrWhiteSpace(name))
+                    {
+                        this.CachedObjects.Add(item, null);
+                        continue;
+                    }
+
+                    managedChest = new ManagedChest(chest, this.GetChestData(name), name);
+                    this.CachedObjects.Add(chest, managedChest);
+                }
+
+                if (managedChest is not null)
+                {
+                    yield return managedChest;
+                }
+            }
+        }
     }
 
     private IEnumerable<GameLocation> AccessibleLocations
@@ -79,6 +154,11 @@ internal class ManagedChests : IModService
     private AssetHandler Assets
     {
         get => this._assetHandler.Value;
+    }
+
+    private IDictionary<Item, IManagedChest> CachedObjects
+    {
+        get => this._cachedObjects.Value;
     }
 
     private IDictionary<string, IChestData> ChestConfigs { get; } = new Dictionary<string, IChestData>();
@@ -101,7 +181,12 @@ internal class ManagedChests : IModService
             return false;
         }
 
-        foreach (var (_, playerChest) in this.PlayerChests)
+        if (this.CachedObjects.TryGetValue(chest, out managedChest))
+        {
+            return managedChest is not null;
+        }
+
+        foreach (var playerChest in this.PlayerChests)
         {
             if (ReferenceEquals(playerChest.Chest, chest))
             {
@@ -110,71 +195,17 @@ internal class ManagedChests : IModService
             }
         }
 
-        foreach (var (placedChest, lazyManagedChest) in this.PlacedChests)
+        foreach (var (_, placedChest) in this.PlacedChests)
         {
-            if (placedChest.ToChest(out var otherChest) && ReferenceEquals(otherChest, chest))
+            if (ReferenceEquals(placedChest.Chest, chest))
             {
-                managedChest = lazyManagedChest.Value;
-                return managedChest is not null;
+                managedChest = placedChest;
+                return true;
             }
         }
 
         managedChest = default;
         return false;
-    }
-
-    private void AddPlacedChest(string locationName, int x, int y, string chestName)
-    {
-        this.AddPlacedChest(new(locationName, x, y, chestName));
-    }
-
-    private void AddPlacedChest(PlacedChest placedChest)
-    {
-        Log.Trace($"Placed Chest added to {placedChest.LocationName} at ({placedChest.X.ToString()}, {placedChest.Y.ToString()}).");
-        if (!this._placedChests.Value.ContainsKey(placedChest))
-        {
-            this._placedChests.Value.Add(
-                placedChest,
-                new(() =>
-                {
-                    if (!placedChest.ToChest(out var chest))
-                    {
-                        return null;
-                    }
-
-                    var config = this.GetChestData(placedChest.ChestName);
-                    return new ManagedChest(chest, config);
-                }));
-        }
-    }
-
-    private void AddPlayerChest(Farmer player, int index)
-    {
-        string name;
-        Chest chest;
-        switch (player.Items[index])
-        {
-            case Chest playerChest when playerChest.IsPlayerChest():
-                chest = playerChest;
-                name = this.Assets.Craftables.SingleOrDefault(info => info.Key == chest.ParentSheetIndex).Value?[0];
-                break;
-            case SObject { heldObject.Value: Chest heldChest } heldObj when heldChest.IsPlayerChest():
-                chest = heldChest;
-                name = this.Assets.Craftables.SingleOrDefault(info => info.Key == heldObj.ParentSheetIndex).Value?[0];
-                break;
-            default:
-                return;
-        }
-
-        Log.Trace($"Player Chest {name} added to {player.Name} in slot {index.ToString()}.");
-        if (string.IsNullOrWhiteSpace(name))
-        {
-            Log.Error($"Tried adding an unknown chest {name}");
-            return;
-        }
-
-        var config = this.GetChestData(name);
-        this._playerChests.Value.Add(new(Game1.player, index), new ManagedChest(chest, config));
     }
 
     private IChestData GetChestData(string name)
@@ -191,161 +222,5 @@ internal class ManagedChests : IModService
         }
 
         return config;
-    }
-
-    private void OnInventoryChanged(object sender, InventoryChangedEventArgs e)
-    {
-        if (!e.IsLocalPlayer)
-        {
-            return;
-        }
-
-        this._playerChests.Value.Clear();
-        for (var index = 0; index < e.Player.Items.Count; index++)
-        {
-            this.AddPlayerChest(e.Player, index);
-        }
-    }
-
-    private void OnModMessageReceived(object sender, ModMessageReceivedEventArgs e)
-    {
-        if (e.FromModID != BetterChests.ModUniqueId)
-        {
-            return;
-        }
-
-        HashSet<PlacedChest> placedChests;
-
-        switch (e.Type)
-        {
-            case "InitPlacedChests":
-                placedChests = e.ReadAs<HashSet<PlacedChest>>();
-                this._placedChests.Value.Clear();
-                foreach (var placedChest in placedChests)
-                {
-                    this.AddPlacedChest(placedChest);
-                }
-
-                break;
-            case "AddPlacedChests":
-                placedChests = e.ReadAs<HashSet<PlacedChest>>();
-                foreach (var placedChest in placedChests)
-                {
-                    this.AddPlacedChest(placedChest);
-                }
-
-                break;
-            case "RemovePlacedChests":
-                placedChests = e.ReadAs<HashSet<PlacedChest>>();
-                foreach (var placedChest in placedChests)
-                {
-                    Log.Trace($"Placed Chest removed from {placedChest.LocationName} at ({placedChest.X.ToString()}, {placedChest.Y.ToString()}).");
-                    this._placedChests.Value.Remove(placedChest);
-                }
-
-                break;
-        }
-    }
-
-    private void OnObjectListChanged(object sender, ObjectListChangedEventArgs e)
-    {
-        var removedChests = new HashSet<PlacedChest>();
-        foreach (var ((x, y), value) in e.Removed)
-        {
-            if (value is Chest placedChest && placedChest.IsPlayerChest())
-            {
-                var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == placedChest.ParentSheetIndex).Value?[0];
-                removedChests.Add(new(e.Location.NameOrUniqueName, (int)x, (int)y, name));
-            }
-
-            if (value is { heldObject.Value: Chest heldChest } && heldChest.IsPlayerChest())
-            {
-                var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == value.ParentSheetIndex).Value?[0];
-                removedChests.Add(new(e.Location.NameOrUniqueName, (int)x, (int)y, name));
-            }
-        }
-
-        foreach (var placedChest in removedChests)
-        {
-            Log.Trace($"Placed Chest removed from {placedChest.LocationName} at ({placedChest.X.ToString()}, {placedChest.Y.ToString()}).");
-            this._placedChests.Value.Remove(placedChest);
-        }
-
-        var addedChests = new HashSet<PlacedChest>();
-        foreach (var ((x, y), value) in e.Added)
-        {
-            if (value is Chest placedChest && placedChest.IsPlayerChest())
-            {
-                var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == placedChest.ParentSheetIndex).Value?[0];
-                addedChests.Add(new(e.Location.NameOrUniqueName, (int)x, (int)y, name));
-            }
-
-            if (value is { heldObject.Value: Chest heldChest } && heldChest.IsPlayerChest())
-            {
-                var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == value.ParentSheetIndex).Value?[0];
-                addedChests.Add(new(e.Location.NameOrUniqueName, (int)x, (int)y, name));
-            }
-        }
-
-        foreach (var placedChest in addedChests)
-        {
-            this.AddPlacedChest(placedChest);
-        }
-
-        if (Context.IsMainPlayer)
-        {
-            this.Helper.Multiplayer.SendMessage(removedChests, "RemovePlacedChests", new[] { BetterChests.ModUniqueId });
-            this.Helper.Multiplayer.SendMessage(addedChests, "AddPlacedChests", new[] { BetterChests.ModUniqueId });
-        }
-    }
-
-    private void OnPeerConnected(object sender, PeerConnectedEventArgs e)
-    {
-        if (e.Peer.IsHost)
-        {
-            return;
-        }
-
-        var placedChests = new HashSet<PlacedChest>(this.PlacedChests.Keys);
-        this.Helper.Multiplayer.SendMessage(placedChests, "InitPlacedChests", new[] { BetterChests.ModUniqueId });
-    }
-
-    private void OnSaveLoaded(object sender, SaveLoadedEventArgs e)
-    {
-        // Initialize Player Items
-        for (var index = 0; index < Game1.player.Items.Count; index++)
-        {
-            this.AddPlayerChest(Game1.player, index);
-        }
-
-        // Initialize Placed Objects
-        foreach (var location in this.AccessibleLocations)
-        {
-            foreach (var ((x, y), obj) in location.Objects.Pairs)
-            {
-                if (obj is Chest chest && chest.IsPlayerChest())
-                {
-                    var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == chest.ParentSheetIndex).Value?[0];
-                    this.AddPlacedChest(location.NameOrUniqueName, (int)x, (int)y, name);
-                    continue;
-                }
-
-                if (obj is { heldObject.Value: Chest heldChest } && heldChest.IsPlayerChest())
-                {
-                    var name = this.Assets.Craftables.SingleOrDefault(info => info.Key == obj.ParentSheetIndex).Value?[0];
-                    this.AddPlacedChest(location.NameOrUniqueName, (int)x, (int)y, name);
-                }
-            }
-
-            if (location is FarmHouse farmHouse && farmHouse.fridge.Value is { } && !farmHouse.fridgePosition.ToVector2().Equals(Vector2.Zero))
-            {
-                this.AddPlacedChest(location.NameOrUniqueName, farmHouse.fridgePosition.X, farmHouse.fridgePosition.Y, "Fridge");
-            }
-        }
-
-        if (Context.IsMainPlayer)
-        {
-            this.Helper.Multiplayer.SendMessage(this.PlacedChests.Keys, "InitPlacedChests", new[] { BetterChests.ModUniqueId });
-        }
     }
 }