+++ BetterChests/CHANGELOG.md
@@ -1,5 +1,9 @@
 ﻿# Better Chests Change Log
 
+## 1.1.0 (February 5, 2022)
+
+* Updated method for keeping Chests in sync for multiplayer
+
 ## 1.0.0 (February 3, 2022)
 
 ### Added