+++ FuryCore/CHANGELOG.md
@@ -1,5 +1,15 @@
 ﻿# FuryCore Change Log
 
+## 1.1.0 (February 5, 2022)
+
+* Added new ICustomTags service
+* Added GMCM integration for new config options
+    * Option to add some custom context tags (enabled by default)
+        * `category_artifact` for items that are Artifact
+        * `category_furniture` for items that are Furniture
+        * `donate_bundle` for items that can be donated to a Community Center bundle
+        * `donate_museum` for items that can be donated to the Museum
+
 ## 1.0.0 (February 3, 2022)
 
 ### Added